+++ cppcore/include/system/SystemModifiers.hpp
@@ -14,8 +14,8 @@ public:
     virtual ~SiteStateModifierImpl() = default;
     
     /// Modify the state (valid or invalid) of sites at the given coordinates
-    virtual void apply(ArrayX<bool>& site_state, CartesianArray const& positions,
-                       ArrayX<sub_id> const& sublattices) const = 0;
+    virtual void apply(ArrayX<bool>& state, CartesianArray const& position,
+                       SubIdRef sublattice) const = 0;
 };
 
 class PositionModifierImpl {
@@ -23,7 +23,7 @@ public:
     virtual ~PositionModifierImpl() = default;
 
     /// Modify the positions system sites
-    virtual void apply(CartesianArray& positions, ArrayX<sub_id> const& sublattices) const = 0;
+    virtual void apply(CartesianArray& position, SubIdRef sublattice) const = 0;
 };
 
 using SiteStateModifier = std::shared_ptr<SiteStateModifierImpl const>;